one_day_servant/requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
mysql-connector-python